requirements.txt
einops==0.8.0
h5py==3.11.0
matplotlib==3.7.5
networkx==3.1
numpy==1.24.3
pandas==2.0.3
pillow==10.4.0
PyWavelets==1.4.1
PyYAML==6.0.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
tensorboard==2.13.0
tensorflow==2.13.1
torch==2.1.1
tqdm==4.66.5
ipykernel==6.29.0
